Wrote more functional question timer system
There are still issues however. The code is not ideal, it uses componentWillReceiveProps
to bypass the fact that the child props can't update (for a new question time).
This timer also is not synced up with the next button.

I'm thinking of moving the timer functions into the parent "Test" component,
while making the TestTimer component solely for rendering html.