Prevent next question if nothing selected
The next question button does not render if the user hasn't
entered picked an answer. I wanted to use a Window.confirm, but
it pauses the entire app, preventing the timer from ticking down.
The questions also deselect themselves when going to a new question.

Also changed the answer array into a JSON format when we send
it to the backend.